refactor : structure
