+++ src/main/java/com/clonecoding/pinterest/domain/pin/service/PinService.java
@@ -1,0 +1,7 @@
+package com.clonecoding.pinterest.domain.pin.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PinService {
+}
