+++ src/main/java/com/clonecoding/pinterest/domain/user/dto/UserLoginDto.java
@@ -1,0 +1,9 @@
+package com.clonecoding.pinterest.domain.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UserLoginDto {
+    private String email;
+    private String password;
+}
