+++ src/main/java/com/clonecoding/pinterest/domain/user/entity/UserRoleEnum.java
@@ -1,0 +1,21 @@
+package com.clonecoding.pinterest.domain.user.entity;
+
+public enum UserRoleEnum {
+    USER(Authority.user),
+    ADMIN(Authority.master);
+
+    private final String authority;
+
+    UserRoleEnum(String authority) {
+        this.authority = authority;
+    }
+
+    public String getAuthority() {
+        return this.authority;
+    }
+
+    public static class Authority {
+        public static final String user = "ROLE_USER";
+        public static final String master = "ROLE_MASTER";
+    }
+}
