+++ src/main/java/com/clonecoding/pinterest/domain/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.clonecoding.pinterest.domain.user.repository;
+
+import com.clonecoding.pinterest.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
