--- src/main/java/com/clonecoding/pinterest/exception/dto/ExceptionResponseDto.java
@@ -1,14 +1,0 @@
-package com.clonecoding.pinterest.exception.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public class ExceptionResponseDto {
-    private HttpStatus httpStatus;
-    private String message;
-}
