--- src/main/java/com/clonecoding/pinterest/pin/entity/CommentLike.java
@@ -1,21 +1,0 @@
-package com.clonecoding.pinterest.pin.entity;
-
-import com.clonecoding.pinterest.user.entity.User;
-import jakarta.persistence.*;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-public class CommentLike {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-}
