--- src/main/java/com/clonecoding/pinterest/pin/service/PinService.java
@@ -1,7 +1,0 @@
-package com.clonecoding.pinterest.pin.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class PinService {
-}
