--- src/main/java/com/clonecoding/pinterest/user/dto/UserLoginDto.java
@@ -1,9 +1,0 @@
-package com.clonecoding.pinterest.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserLoginDto {
-    private String email;
-    private String password;
-}
