--- src/main/java/com/clonecoding/pinterest/user/entity/User.java
@@ -1,39 +1,0 @@
-package com.clonecoding.pinterest.user.entity;
-
-import com.clonecoding.pinterest.entity.TimeStamped;
-import com.clonecoding.pinterest.user.dto.UserCreateRequestDto;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class User extends TimeStamped {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String username;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column
-    private String profileImageUrl;
-
-    @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private UserRoleEnum role;
-
-    public User(UserCreateRequestDto requestDto){
-        this.email = requestDto.getEmail();
-        this.username = requestDto.getUsername();
-        this.password = requestDto.getPassword();
-        this.role = requestDto.getRole();
-    }
-}
