--- src/main/java/com/clonecoding/pinterest/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.clonecoding.pinterest.user.repository;
-
-import com.clonecoding.pinterest.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
